src/mcp/oracle/pyproject.toml
[project]
name = "ovd-mcp-oracle"
version = "0.1.0"
description = "OVD Platform — MCP Server Oracle multi-sede (CAS/CAT/CAV)"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.3.0",
    "oracledb>=2.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
]

[tool.uv]
